+++ api/test/unit/services/user.service.spec.ts
@@ -1,5 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ConfigService } from '@nestjs/config';
+import { Request } from 'express';
 import { PrismaService } from '../../../src/services/prisma.service';
 import { UserService } from '../../../src/services/user.service';
 import { randomUUID } from 'crypto';
@@ -15,6 +16,7 @@ import { SendGridService } from '../../../src/services/sendgrid.service';
 import { User } from '../../../src/dtos/users/user.dto';
 import { PermissionService } from '../../../src/services/permission.service';
 import { permissionActions } from '../../../src/enums/permissions/permission-actions-enum';
+import { OrderByEnum } from '../../../src/enums/shared/order-by-enum';
 
 describe('Testing user service', () => {
   let service: UserService;
@@ -1404,6 +1406,7 @@ describe('Testing user service', () => {
           id: 'requestingUser id',
           userRoles: { isAdmin: true },
         } as unknown as User,
+        { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
       );
       expect(prisma.userAccounts.findUnique).toHaveBeenCalledWith({
         include: {
@@ -1475,6 +1478,7 @@ describe('Testing user service', () => {
           id: 'requestingUser id',
           userRoles: { isAdmin: true },
         } as unknown as User,
+        { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
       );
       expect(prisma.userAccounts.findUnique).toHaveBeenCalledWith({
         include: {
@@ -1555,6 +1559,7 @@ describe('Testing user service', () => {
               id: 'requestingUser id',
               userRoles: { isAdmin: true },
             } as unknown as User,
+            { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
           ),
       ).rejects.toThrowError('emailInUse');
       expect(prisma.userAccounts.findUnique).toHaveBeenCalledWith({
@@ -1616,6 +1621,7 @@ describe('Testing user service', () => {
           id: 'requestingUser id',
           userRoles: { isAdmin: true },
         } as unknown as User,
+        { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
       );
       expect(prisma.userAccounts.findUnique).toHaveBeenCalledWith({
         include: {
@@ -1784,5 +1790,220 @@ describe('Testing user service', () => {
       );
       expect(res).toEqual(false);
     });
+  });
+
+  it('should request single use code but user does not exist', async () => {
+    const id = randomUUID();
+    emailService.sendSingleUseCode = jest.fn();
+    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue(null);
+    prisma.userAccounts.update = jest.fn().mockResolvedValue({
+      id,
+    });
+
+    const res = await service.requestSingleUseCode(
+      {
+        email: 'example@exygy.com',
+      },
+      { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
+    );
+
+    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
+      where: {
+        email: 'example@exygy.com',
+      },
+      include: {
+        jurisdictions: true,
+      },
+    });
+    expect(prisma.userAccounts.update).not.toHaveBeenCalled();
+    expect(emailService.sendSingleUseCode).not.toHaveBeenCalled();
+    expect(res).toEqual({
+      success: true,
+    });
+  });
+
+  it('should request single use code but jurisdiction does not exist', async () => {
+    const id = randomUUID();
+    emailService.sendSingleUseCode = jest.fn();
+    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue({
+      id,
+    });
+    prisma.jurisdictions.findFirst = jest.fn().mockResolvedValue(null);
+    prisma.userAccounts.update = jest.fn().mockResolvedValue({
+      id,
+    });
+
+    await expect(
+      async () =>
+        await service.requestSingleUseCode(
+          {
+            email: 'example@exygy.com',
+          },
+          { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
+        ),
+    ).rejects.toThrowError('Jurisidiction juris 1 does not exists');
+
+    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
+      where: {
+        email: 'example@exygy.com',
+      },
+      include: {
+        jurisdictions: true,
+      },
+    });
+    expect(prisma.jurisdictions.findFirst).toHaveBeenCalledWith({
+      select: {
+        id: true,
+        allowSingleUseCodeLogin: true,
+      },
+      where: {
+        name: 'juris 1',
+      },
+      orderBy: {
+        allowSingleUseCodeLogin: OrderByEnum.DESC,
+      },
+    });
+    expect(prisma.userAccounts.update).not.toHaveBeenCalled();
+    expect(emailService.sendSingleUseCode).not.toHaveBeenCalled();
+  });
+
+  it('should request single use code but jurisdiction disallows single use code login', async () => {
+    const id = randomUUID();
+    emailService.sendSingleUseCode = jest.fn();
+    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue({
+      id,
+    });
+    prisma.jurisdictions.findFirst = jest.fn().mockResolvedValue({
+      id: randomUUID(),
+      allowSingleUseCodeLogin: false,
+    });
+    prisma.userAccounts.update = jest.fn().mockResolvedValue({
+      id,
+    });
+
+    await expect(
+      async () =>
+        await service.requestSingleUseCode(
+          {
+            email: 'example@exygy.com',
+          },
+          { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
+        ),
+    ).rejects.toThrowError('Single use code login is not setup for juris 1');
+
+    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
+      where: {
+        email: 'example@exygy.com',
+      },
+      include: {
+        jurisdictions: true,
+      },
+    });
+    expect(prisma.jurisdictions.findFirst).toHaveBeenCalledWith({
+      select: {
+        id: true,
+        allowSingleUseCodeLogin: true,
+      },
+      where: {
+        name: 'juris 1',
+      },
+      orderBy: {
+        allowSingleUseCodeLogin: OrderByEnum.DESC,
+      },
+    });
+    expect(prisma.userAccounts.update).not.toHaveBeenCalled();
+    expect(emailService.sendSingleUseCode).not.toHaveBeenCalled();
+  });
+
+  it('should request single use code but jurisdictionname was not sent', async () => {
+    const id = randomUUID();
+    emailService.sendSingleUseCode = jest.fn();
+    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue({
+      id,
+    });
+    prisma.jurisdictions.findFirst = jest.fn().mockResolvedValue({
+      id,
+    });
+    prisma.userAccounts.update = jest.fn().mockResolvedValue({
+      id,
+    });
+
+    await expect(
+      async () =>
+        await service.requestSingleUseCode(
+          {
+            email: 'example@exygy.com',
+          },
+          {} as unknown as Request,
+        ),
+    ).rejects.toThrowError(
+      'jurisdictionname is missing from the request headers',
+    );
+
+    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
+      where: {
+        email: 'example@exygy.com',
+      },
+      include: {
+        jurisdictions: true,
+      },
+    });
+    expect(prisma.jurisdictions.findFirst).not.toHaveBeenCalled();
+    expect(prisma.userAccounts.update).not.toHaveBeenCalled();
+    expect(emailService.sendSingleUseCode).not.toHaveBeenCalled();
+  });
+
+  it('should successfully request single use code', async () => {
+    const id = randomUUID();
+    emailService.sendSingleUseCode = jest.fn();
+    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue({
+      id,
+    });
+    prisma.jurisdictions.findFirst = jest.fn().mockResolvedValue({
+      id,
+      allowSingleUseCodeLogin: true,
+    });
+    prisma.userAccounts.update = jest.fn().mockResolvedValue({
+      id,
+    });
+
+    const res = await service.requestSingleUseCode(
+      {
+        email: 'example@exygy.com',
+      },
+      { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
+    );
+
+    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
+      where: {
+        email: 'example@exygy.com',
+      },
+      include: {
+        jurisdictions: true,
+      },
+    });
+    expect(prisma.jurisdictions.findFirst).toHaveBeenCalledWith({
+      select: {
+        id: true,
+        allowSingleUseCodeLogin: true,
+      },
+      where: {
+        name: 'juris 1',
+      },
+      orderBy: {
+        allowSingleUseCodeLogin: OrderByEnum.DESC,
+      },
+    });
+    expect(prisma.userAccounts.update).toHaveBeenCalledWith({
+      data: {
+        singleUseCode: expect.anything(),
+        singleUseCodeUpdatedAt: expect.anything(),
+      },
+      where: {
+        id,
+      },
+    });
+    expect(emailService.sendSingleUseCode).toHaveBeenCalled();
+    expect(res.success).toEqual(true);
   });
 });
